[Activity Feed] Moving new files into Activity directory
