+++ components/Activity/ContributionAmount.tsx
